Added x265 as an external
Added the externals\x265\CMakeLists.txt which contains the name (x265) of the project and adds the source/ folder as a subdirectory since the CMakeLists.txt of the x265 git repository is not in the top level but in the .\source\ folder. The externals\x265\CMakeLists.txt will be copied into the downloaded  x265 folder.
Patched the externals\CMakeExternals.cmake to include the x265 library.
These changes are so far not tested under Linux.