Implemented CPU encoding using x264 and x265
Added find_package for x264 to the externals.
Changed the configuration of the build command used to build the x265 external.
Added the Latency measurement class.
Added the AbstractVideoEncoder which is the parent class used by the CpuVideoEncoder and in future by the GpuVideoEncoder.
The CpuVideoEncoder uses x264 or x265 to encode the content of a view. This does (so far) not include the UI.