MPI implementation and minor eincoder fixes
Udpated the encoder plugin readme as there was one step missing in the build instructions for x264.
Made the Network Plugin optional for the Encoder Plugin.
Added all parameters to the unchangeable callback to disallow changes while the encoder is running.
Implemented the MPI input/output for the Encoder Plugin.
Added the MpiIntraCall and the MpiIntraCommunicator in order to allow communication via MPI.